Fix button creation logic
Button creation logic was able to be simplified. Previously there
was a precondition in which an Item's ID and its row in the
ItemNumbering.csv needed to be equal as the row index value was
used to "find" the icon's file name.

The item's ID should be equal to the row index anyway. So we're
removing the outer loop and relying on the Item ID.

This also fixes an issue with adding new items.