Setup logging according to the logging cookbook
Python has the logging cookbook (please see
https://docs.python.org/3/library/logging.html for more info). It
basically allows you tos setup the logger to log to both a file
and the console, while using two different formatters.

So converted the logging functionality to using the python
logging module with two seperate handlers, one for console
and one for file.

This also allowed us to clean up the print statements that
duplicated the log messages as the handlers would handle
console and file output.

Also setup the formatters to capture the time of the log message
so we don't need to record the time in the log messasge anymore.